+++ packages/react/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## [4.44.0](https://github.com/factorialco/f0/compare/f0-react-v4.43.0...f0-react-v4.44.0) (2026-07-20)
+
+
+### Features
+
+* **F0PdfViewer:** document viewer kinds (sheet, docx, text) with toolbar actions ([#4741](https://github.com/factorialco/f0/issues/4741)) ([ddbeb67](https://github.com/factorialco/f0/commit/ddbeb67fa5b0157479d77251dee8c0d2534f112c))
+
+## [4.43.0](https://github.com/factorialco/f0/compare/f0-react-v4.42.1...f0-react-v4.43.0) (2026-07-20)
+
+
+### Features
+
+* **OneCalendar:** month and year dropdowns in the DatePicker calendar header ([#4736](https://github.com/factorialco/f0/issues/4736)) ([24d6a64](https://github.com/factorialco/f0/commit/24d6a64c88a8abc7cde598b54d13193f094d895f))
+
+## [4.42.1](https://github.com/factorialco/f0/compare/f0-react-v4.42.0...f0-react-v4.42.1) (2026-07-20)
+
+
+### Bug Fixes
+
+* **F0Card:** make whole card clickable except footer ([#4745](https://github.com/factorialco/f0/issues/4745)) ([a23066e](https://github.com/factorialco/f0/commit/a23066e2badb25ba47b10049f123e6373e1c5a1e))
+
+## [4.42.0](https://github.com/factorialco/f0/compare/f0-react-v4.41.1...f0-react-v4.42.0) (2026-07-17)
+
+
+### Features
+
+* **storybook:** show component status + DoD checklist on docs pages ([#4724](https://github.com/factorialco/f0/issues/4724)) ([f46f4ae](https://github.com/factorialco/f0/commit/f46f4ae2dc15cf73ce15daffd0d9393fb5d5ca1f))
+
+## [4.41.1](https://github.com/factorialco/f0/compare/f0-react-v4.41.0...f0-react-v4.41.1) (2026-07-17)
+
+
+### Bug Fixes
+
+* **OneDateNavigator:** revive persisted string dates before use ([#4733](https://github.com/factorialco/f0/issues/4733)) ([6d6fcbe](https://github.com/factorialco/f0/commit/6d6fcbe2be79f4aefce71c1b8e2e14a3fa1e1075))
+
 ## [4.41.0](https://github.com/factorialco/f0/compare/f0-react-v4.40.1...f0-react-v4.41.0) (2026-07-17)